% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{canopy.plottree.mod}
\alias{canopy.plottree.mod}
\title{Split confCI character}
\usage{
canopy.plottree.mod(tree, pdf = NULL, pdf.name = NULL, txt = NULL,
  txt.name = NULL, save = FALSE, rdata.name = NULL)
}
\value{
tree, mutation list
}
\description{
Split confCI character
}
