% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-capony-par.R
\name{canopy.sample.parallel.mod}
\alias{canopy.sample.parallel.mod}
\title{Run Canopy in parallel}
\usage{
canopy.sample.parallel.mod(
  R,
  X,
  WM,
  Wm,
  epsilonM,
  epsilonm,
  C = NULL,
  Y,
  K,
  numchain,
  max.simrun,
  min.simrun,
  writeskip,
  projectname,
  cell.line = NULL,
  plot.likelihood = NULL,
  nCores = 1
)
}
\arguments{
\item{Y}{An output object from the make_Y_mat() function.}

\item{K}{Vector of the number of clones to be used.}

\item{numchain}{Number of chains for the MCMC. Default is 15.}

\item{max.simrun}{Maximum number of simulations to perform before convergence. Default is 100000.}

\item{min.simrun}{Minimum number of simulations to perform before convergence. Default is 10000.}

\item{projectname}{Name of the project to save results.}

\item{sna_obj}{An output object from the make_sna_mat() function.}

\item{cna_obj}{An output object from the make_cna_mat() function.}

\item{burnin}{Burn-in period for the MCMC. Default is 10.}

\item{thin}{Thining parameter for the MCMC. Default is 5.}

\item{parallel}{Should this be run in parallel. Default is FALSE.}
}
\value{
Y
}
\description{
Run Canopy in parallel
}
