% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottree-mod.R
\name{canopy_plottree_mod}
\alias{canopy_plottree_mod}
\title{canopy.plottree.mod}
\usage{
canopy_plottree_mod(
  tree,
  pdf = NULL,
  pdf.name = NULL,
  txt = NULL,
  txt.name = NULL,
  save = FALSE,
  rdata.name = NULL,
  cex_names = NULL
)
}
\value{
tree, mutation list
}
\description{
Split confCI character
}
\examples{
\dontrun{
canopy.plottree.mod(tree, pdf = NULL, pdf.name = NULL, txt = NULL, txt.name = NULL,save = FALSE, rdata.name = NULL)
}
}
