% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{computeCCF}
\alias{computeCCF}
\title{Compute CCF and multiplicity issue in major and minor CN.}
\usage{
computeCCF(vaf, tt, minor, purity, multiplicity = NULL)
}
\value{
ccf, rawccf, out, multiplicity
}
\description{
Compute CCF and multiplicity issue in major and minor CN.
}
