% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germline_filter.R
\name{germline_filter}
\alias{germline_filter}
\title{Plot expected germline and somatic VAF along with observed somatic VAF.
And retrieve the the mutational status infered by the method.
Variants found to have an observed VAF below 2 standard deviations below the expected value are considered "subclonal".
Those falling in the +/- 2 standard deviation range from the expected VAF are annoatated "neutral".
Those with VAF greater than 2 starndard deviation above the expected value are annotated as "artifacts".}
\usage{
germline_filter(mat, purity, alpha = 0.05, plot = TRUE, sample)
}
\value{
mat.out
}
\description{
Plot expected germline and somatic VAF along with observed somatic VAF.
And retrieve the the mutational status infered by the method.
Variants found to have an observed VAF below 2 standard deviations below the expected value are considered "subclonal".
Those falling in the +/- 2 standard deviation range from the expected VAF are annoatated "neutral".
Those with VAF greater than 2 starndard deviation above the expected value are annotated as "artifacts".
}
