% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{get_confBound}
\alias{get_confBound}
\title{Split confCI character}
\usage{
get_confBound(x)
}
\value{
xx.l, xx.u
}
\description{
Split confCI character
}
\examples{
\dontrun{
computeCCF(vaf, tt, minor, purity, multiplicity=NULL)
}
}
