% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{get_expectedVAF}
\alias{get_expectedVAF}
\title{Calculate expected germline and somatic VAFs along with their CI given tcn and multiplicity}
\usage{
get_expectedVAF(multiplicity, tcn, purity, alt, total, alpha = 0.05)
}
\value{
ccf, multiplicity, clonal_status, ccfCI
}
\description{
Calculate expected germline and somatic VAFs along with their CI given tcn and multiplicity
}
