% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.R
\name{get_results_canopy}
\alias{get_results_canopy}
\title{get_results}
\usage{
get_results_canopy(cna.obj, tree, projectname, path = ".")
}
\arguments{
\item{projectname}{Name of the project to save results.}

\item{path}{Relative path to where results should be saved. Default is current directory.}
}
\value{
Results pdf
}
\description{
Get the results from a canopy optimized run
}
