% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cna.R
\name{make_cna_mat}
\alias{make_cna_mat}
\title{Create a Canopy ready CNA matrix and their corresponding variance matrices}
\usage{
make_cna_mat(
  cna.files = list.files(),
  path = ".",
  sample.names = NULL,
  cval = 100,
  snp.nbhd = 250,
  epsilon = 0
)
}
\arguments{
\item{cna.files}{Names of the files to be read in and processed. Default will be all the files found in the working directory.}

\item{path}{relative path to the folder containing the samples to be processed. Default is current directory.}

\item{sample.names}{Custom names to be given to the samples. Default is NULL.}

\item{cval}{Default is 100.}

\item{snp.nbhd}{Default is 250.}

\item{epsilon}{Default is 0.}
}
\value{
WM, Wm,epsM, epsm
}
\description{
Create a Canopy ready CNA matrix and their corresponding variance matrices
}
