% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sna.R
\name{make_sna_mat}
\alias{make_sna_mat}
\title{Create a Canopy ready SNA matrices}
\usage{
make_sna_mat(mut.files = list.files(), path = ".",
  sample.names = NULL)
}
\arguments{
\item{mut.files}{Names of the files to be read in and processed. Default will be all the files found in the working directory.}

\item{path}{relative path to the folder containing the samples to be processed. Default is current directory.}

\item{sample.names}{Custom names to be given to the samples. Default is NULL.}
}
\value{
X, R
}
\description{
Create a Canopy ready SNA matrices
}
