% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_int.R
\name{make_Y_mat}
\alias{make_Y_mat}
\title{Create a Canopy ready Y sna-cna interaction matrix}
\usage{
make_Y_mat(sna_obj, cna_obj)
}
\arguments{
\item{sna_obj}{An output object from the make_sna_mat() function.}

\item{cna_obj}{An output object from the make_cna_mat() function.}
}
\value{
Y
}
\description{
Create a Canopy ready Y sna-cna interaction matrix
}
