% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-phylo.R
\name{opt_phylo}
\alias{opt_phylo}
\title{Get optimal phylogeny}
\usage{
opt_phylo(
  sampchain,
  projectname,
  K,
  numchain = 15,
  max.simrun = 1e+05,
  min.simrun = 10000,
  path = ".",
  burnin = 10,
  thin = 5
)
}
\arguments{
\item{sampchain}{An output object from the run_canpy() function.}

\item{projectname}{Name of the project to save results.}

\item{K}{Vector of the number of clones to be used.}

\item{numchain}{Number of chains for the MCMC. Default is 15.}

\item{max.simrun}{Maximum number of simulations to perform before convergence. Default is 15000.}

\item{min.simrun}{Minimum number of simulations to perform before convergence. Default is 5000.}

\item{path}{Relative path to where results should be saved. Default is current directory.}

\item{burnin}{Burn-in period for the MCMC. Default is 10.}

\item{thin}{Thining parameter for the MCMC. Default is 5.}
}
\value{
Optimal phylogeny.
}
\description{
Get optimal phylogeny
}
