% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-can.R
\name{plot_cna}
\alias{plot_cna}
\title{Plot CNA heatmap}
\usage{
plot_cna(dat, patients = NULL, ordered = NULL, outcome = NULL)
}
\arguments{
\item{dat}{Data}

\item{patients}{patients}

\item{ordered}{ordered}

\item{outcome}{outcome}
}
\value{
facets heatmap
}
\description{
Plot CNA heatmap
}
