% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TumorEvolution-package.R
\docType{package}
\name{TumorEvolution-package}
\alias{TumorEvolution}
\alias{TumorEvolution-package}
\title{TumorEvolution: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
