% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{confCCF}
\alias{confCCF}
\title{Confidence interval for CCF at a default alpha of 0.05 or 95% CI}
\usage{
confCCF(alt, ref, tt, minor, purity, multiplicity, alpha = 0.05)
}
\value{
lower, upper
}
\description{
Confidence interval for CCF at a default alpha of 0.05 or 95% CI
}
