% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{get_ccf}
\alias{get_ccf}
\title{Wrapper function to get necessary things from CCF}
\usage{
get_ccf(alt, total, tcn, lcn, purity)
}
\arguments{
\item{alt}{alt}

\item{total}{total}

\item{tcn}{tcn}

\item{lcn}{lcn}

\item{purity}{purity}
}
\value{
ccf, multiplicity, clonal_status, ccfCI
}
\description{
Wrapper function to get necessary things from CCF
}
\examples{
\dontrun{
computeCCF(vaf, tt, minor, purity, multiplicity=NULL)
}
}
