% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.R
\name{get_results_canopy}
\alias{get_results_canopy}
\title{get_results}
\usage{
get_results_canopy(
  cna.obj,
  tree,
  projectname,
  path = ".",
  path_mut = ".",
  mut.files,
  sample.names = NULL,
  cex_cna = 5,
  epsilon = 0
)
}
\arguments{
\item{projectname}{Name of the project to save results.}

\item{path}{Relative path to where results should be saved. Default is current directory.}

\item{path_mut}{Relative path to the mutation files used to run canopy}

\item{mut.files}{Mutation file names (same order as in the canopy run)}

\item{sample.names}{Sample names for the files}

\item{cex_cna}{size of font in CNA heatmap}

\item{epsilon}{Segment aggregation}
}
\value{
Results pdf
}
\description{
Get the results from a canopy optimized run
}
