% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt-phylo.R
\name{opt_phylo}
\alias{opt_phylo}
\title{Get optimal phylogeny}
\usage{
opt_phylo(
  sampchain,
  projectname,
  K,
  numchain = 15,
  path = ".",
  burnin = 10,
  thin = 5,
  cutoff = 0.05
)
}
\arguments{
\item{sampchain}{An output object from the run_canpy() function.}

\item{projectname}{Name of the project to save results.}

\item{K}{Vector of the number of clones to be used.}

\item{numchain}{Number of chains for the MCMC. Default is 15.}

\item{path}{Relative path to where results should be saved. Default is current directory.}

\item{burnin}{Burn-in period for the MCMC. Default is 10.}

\item{thin}{Thining parameter for the MCMC. Default is 5.}

\item{cutoff}{post configuration cutoff. Default is 0.05}
}
\value{
Optimal phylogeny.
}
\description{
Get optimal phylogeny
}
