% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_canopy.R
\name{run_canopy}
\alias{run_canopy}
\title{Run Canopy}
\usage{
run_canopy(
  sna_obj,
  cna_obj,
  Y,
  projectname,
  K,
  numchain = 15,
  max.simrun = 1e+05,
  min.simrun = 10000,
  burnin = 10,
  thin = 5,
  parallel = FALSE,
  nCores = 1
)
}
\arguments{
\item{sna_obj}{An output object from the make_sna_mat() function.}

\item{cna_obj}{An output object from the make_cna_mat() function.}

\item{Y}{An output object from the make_Y_mat() function.}

\item{projectname}{Name of the project to save results.}

\item{K}{Vector of the number of clones to be used.}

\item{numchain}{Number of chains for the MCMC. Default is 15.}

\item{max.simrun}{Maximum number of simulations to perform before convergence. Default is 100000.}

\item{min.simrun}{Minimum number of simulations to perform before convergence. Default is 10000.}

\item{burnin}{Burn-in period for the MCMC. Default is 10.}

\item{thin}{Thining parameter for the MCMC. Default is 5.}

\item{parallel}{Should this be run in parallel. Default is FALSE.}
}
\value{
Y
}
\description{
Run Canopy
}
